requirements.txt
blinker==1.7.0
cachelib==0.10.2
click==8.1.7
Flask==3.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.3
MarkupSafe==2.1.3
Pillow==10.1.0
PyMuPDF==1.23.6
PyMuPDFb==1.23.6
python-dotenv==1.0.0
reportlab==4.0.7
typing_extensions==4.8.0
Werkzeug==3.0.1
